+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/common/RecordStatus.java
@@ -1,11 +1,10 @@
 package org.nervos.mercury.model.common;
 
-import java.math.BigInteger;
 import org.nervos.mercury.model.resp.AssetStatus;
 
 public class RecordStatus {
 
   public AssetStatus type;
 
-  public BigInteger value;
+  public int value;
 }
