+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/GetBalanceResponse.java
@@ -1,12 +1,11 @@
 package org.nervos.mercury.model.resp;
 
 import com.google.gson.annotations.SerializedName;
-import java.math.BigInteger;
 import java.util.List;
 
 public class GetBalanceResponse {
   public List<BalanceResponse> balances;
 
   @SerializedName("tip_block_number")
-  public BigInteger tipBlockNumber;
+  public int tipBlockNumber;
 }
